+++ workspaces/grafana/plugins/grafana/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-grafana
 
+## 0.3.0
+
+### Minor Changes
+
+- d18e93e: Backstage version bump to v1.35.1
+
 ## 0.2.0
 
 ### Minor Changes