+++ workspaces/opencost/plugins/opencost/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-opencost
 
+## 0.4.0
+
+### Minor Changes
+
+- 034249c: Backstage version bump to v1.35.1
+
 ## 0.3.0
 
 ### Minor Changes