+++ workspaces/stackstorm/plugins/stackstorm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-stackstorm
 
+## 0.3.0
+
+### Minor Changes
+
+- 4cb973d: Backstage version bump to v1.35.1
+
 ## 0.2.0
 
 ### Minor Changes